unite the schema derivation
